몬스터 길찾기 알고리즘 추가(AStar)

+++ GameDataUploader/Model.cs
@@ -7,7 +7,7 @@ namespace GameDataUploader
     {
         public int TileId { get; set; }
         public int MapId { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public short X { get; set; }
+        public short Y { get; set; }
     }
 }
